Improved Particle FX
Made butterflies spawn in a squashed sphere and vortex around the central axis, with some curl randomness and drag towards the central axis.
Also made them spawn continuously with a fade-in and fade-out so that we can modify the spawn rate dynamically during play. Updated the ButterflySpawner actor to adjust the spawnrate during its overlap timeline. Note: The fade was done using mask, i.e. I didn't change the material to transparent. I just added used a previously unused channel of the dynamic parameter to multiply the total mask.
Niagara emitter is a GPU sim now. I think it'll work fine as such, but worst case it can be changed back. Currently unused: There's a node that lets particles sort of avoid collision with other solid meshes. It makes their flight path a bit wack and I don't know if we need it, but it's there. Requires GPU sim.

Anything else: Will require significantly more effort. If we want really intricate swarm behavior we can use boid algorithms but that's like, 10 times more complex and I am not sure this is the project where it'll pay off.